+++ src/database/database_connection.py
@@ -2,11 +2,10 @@ import sqlite3
 
 from utilities.config_manager import ConfigManager
 
-CFG: ConfigManager = ConfigManager()
-CFG.create_config()
-DATABASE_FILE: str = CFG.get_database_path()
-CONNECTION: sqlite3.Connection = sqlite3.connect(DATABASE_FILE)
-
 
 def get_database_connection() -> sqlite3.Connection:
-    return CONNECTION
+    cfg: ConfigManager = ConfigManager()
+    cfg.create_config()
+    database_file: str = cfg.get_database_path()
+    connection: sqlite3.Connection = sqlite3.connect(database_file)
+    return connection